Add support for multi-arch images
Presently, when `incert` encounters a multi-arch image it will resolve
it down to the platform specific manifest (by default linux/amd64) and
insert certificates into that. A single manifest is published to the
destination.

With this change, `incert` will insert the certificates into each image
in the list and publish a new index to the destination, derived from the
source.

I think this matches the user's expectation of how the command would behave and
removes some toil for users who are interested in actually consuming
multiple architectures.

This changes the default behaviour of `incert` when it encounters a
multi-arch image and could be considered a breaking change.

The alternative would be to put this behind a `--recursive` flag or
something but my feeling is that this is the expected default behaviour.

Signed-off-by: Rob Best <rob.best@chainguard.dev>